refactor: update folder structure following feedbacks
