+++ packages/utils/theme/tsconfig.json
@@ -1,10 +1,9 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../../../tsconfig.json",
   "compilerOptions": {
     "baseUrl": ".",
     "rootDir": "./",
      "outDir": "./dist",
-    "module": "node16",
     "declaration": true,
   },
   "include": ["**/*", "../../global.d.ts"]
